+++ mycodo/inputs/examples/bme280_ttn.py
@@ -173,7 +173,7 @@ class InputModule(AbstractInput):
                     self.value_get(2),
                     self.value_get(0))
                 self.lock_acquire(self.lock_file, timeout=10)
-                if self.locked[self.lock_file]:
+                if self.lockfile.locked[self.lock_file]:
                     try:
                         self.serial_send = self.serial.Serial(
                             port=self.serial_device,