+++ mycodo/inputs/sht31_smart_gadget.py
@@ -398,7 +398,7 @@ class InputModule(AbstractInput):
         self.return_dict = copy.deepcopy(measurements_dict)
 
         self.lock_acquire(self.lock_file, timeout=3600)
-        if self.locked[self.lock_file]:
+        if self.lockfile.locked[self.lock_file]:
             self.logger.debug("Starting measurement")
             try:
                 if not self.initialized: